+++ component/backend/View/Users/Html.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * @package   AkeebaLoginGuard
+ * @copyright Copyright (c)2016-2018 Nicholas K. Dionysopoulos / Akeeba Ltd
+ * @license   GNU General Public License version 3, or later
+ */
+
+namespace Akeeba\LoginGuard\Admin\View\Users;
+
+// Protect from unauthorized access
+use FOF30\View\DataView\Html as BaseView;
+
+defined('_JEXEC') or die();
+
+/**
+ * Users view
+ *
+ * @since   3.1.0
+ */
+class Html extends BaseView
+{
+}