+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+# 3.0.1
+
+**New**
+
+* Users page to see which users have 2SV enabled or not
+* Ability to force-disable TSV for specific user groups
+* gh-49 Forced 2SV for specific user groups
+
+**Other changes**
+
+* Joomla! 3.9 backend Components menu item compatibility
+* Allow com_ajax in the captive page (used by cookie banners and similar)
+
+**Bug fixes**
+
+* U2F might fail on Firefox due to a missing semicolon
+
 # 3.0.0
 
 **Other changes**